refactor: use new types & enums
